moved static files to static dir
