+++ static/votecheck.js
